refactor(core-components-amount): case sensetive renaming
