+++ packages/amount/src/component.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta, Story, Props, Title } from '@storybook/addon-docs/blocks';
 import { text, select, boolean, number } from '@storybook/addon-knobs';
 
-import { Amount } from './component';
+import { Amount } from '.';
 import { CURRENCY_CODES } from './utils/currencyCodes';
 import { name, version } from '../package.json';
