+++ packages/amount/src/index.ts
@@ -1,1 +1,1 @@
-export * from './component';
+export * from './с-omponent';
