+++ packages/amount/src/с-omponent.tsx
